+++ docs/changelog.rst
@@ -37,6 +37,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - macOS: When the Apple Color Emoji font lacks an emoji glyph search for it in other
   installed fonts (:iss:`3591`)
 
+- macOS: Fix rendering getting stuck on some machines after sleep/screensaver
+  (:iss:`2016`)
+
 - Add a few more special commandline arguments for the launch command. Now all
   ``KITTY_PIPE_DATA`` is also available via command line argument substitution
   (:iss:`3593`)